drivers:platform:linux:linux_spi: Fix compile warnings
This commit fixes two of the following warnings:

warning: assignment to ‘__u64’ {aka ‘long long unsigned int’} from
‘uint8_t *’ {aka ‘unsigned char *’} makes integer from pointer without a
cast [-Wint-conversion]

Please see: https://github.com/torvalds/linux/blob/6e764bcd1cf72a2846c0e53d3975a09b242c04c9/tools/spi/spidev_test.c#L123

+++ drivers/platform/linux/linux_spi.c
@@ -206,8 +206,8 @@ static int32_t linux_spi_transfer(struct spi_desc *desc,
 		return -ENOMEM;
 
 	for (i = 0; i < len; i++) {
-		tr[i].tx_buf = msgs[i].tx_buff;
-		tr[i].rx_buf = msgs[i].rx_buff;
+		tr[i].tx_buf = (unsigned long) msgs[i].tx_buff;
+		tr[i].rx_buf = (unsigned long) msgs[i].rx_buff;
 		tr[i].len = msgs[i].bytes_number;
 		tr[i].cs_change = msgs[i].cs_change;
 	}
